% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ms2.R
\name{.get_spectra2}
\alias{.get_spectra2}
\alias{.get_spectra2_closeRT}
\title{Extract MS2 from Spectra object.}
\usage{
.get_spectra2(
  sps,
  cpid,
  mzmin,
  mzmax,
  rtmin,
  rtmax,
  spectraOrigin,
  mzdiff = 0.001,
  rtdiff = 0
)

.get_spectra2_closeRT(
  sps,
  cpid,
  rt,
  mzmin,
  mzmax,
  rtmin,
  rtmax,
  spectraOrigin,
  mzdiff = 0.001,
  rtdiff = 0
)
}
\arguments{
\item{sps}{Spectra object.}

\item{cpid}{Compound peak id.}

\item{mzmin}{Minimum value of peak mz.}

\item{mzmax}{Maximum value of peak mz.}

\item{rtmin}{Minimum value of peak rt.}

\item{rtmax}{Maximum value of peak rt.}

\item{spectraOrigin}{Path of spectra.}

\item{mzdiff}{c(mzmin - mzdiff, mzmax + mzdiff) constitues the precursor mz search range.
The mzdiff is determined by the number of decimal places in the spectra precursor mz recording.
Example: precursorMz = 312.83, mzdiff = 0.01.}

\item{rtdiff}{c(rtmin - rtdiff, rtmax + rtdiff) constitues the retention time search range.}

\item{rt}{Retention time of target peak.}
}
\value{
Spectra object.
}
\description{
Extract MS2 for compound peak of data from Spectra object.
}
\examples{
.get_spectra2(sps = sps, cpid = data$peaksInfo[7000, ]$cpid,
              mzmin = data$peaksInfo[7000, ]$mzmin, mzmax = data$peaksInfo[7000, ]$mzmax,
              rtmin = data$peaksInfo[7000, ]$rtmin, rtmax = data$peaksInfo[7000, ]$rtmax,
              spectraOrigin = "D:\\\\fudan\\\\Projects\\\\2024\\\\MultichannelR\\\\Progress\\\\build_package\\\\test_data\\\\AP\\\\mix1\\\\CXYC136A240712-dx-DEANS-AP_mix1_0712-1.mzML")
}
