% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_group.R
\name{assign_plexInfo}
\alias{assign_plexInfo}
\title{Assign plex information}
\usage{
assign_plexInfo(data, plexPara, thread = 1)
}
\arguments{
\item{data}{data list.}

\item{plexPara}{plex parameter list.}

\item{thread}{thread.}
}
\value{
data list.
}
\description{
Assign plex information for each peak containing precursorMz, specialFragment and plexIdx.
}
\examples{
data <- assign_plexInfo(data, plexPara, thread = 3)
}
