% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_group.R
\name{assign_tagNum}
\alias{assign_tagNum}
\title{Assign tagNum}
\usage{
assign_tagNum(data, plexPara, thread = 1)
}
\arguments{
\item{data}{data list.}

\item{plexPara}{plexPara list.}

\item{thread}{thread.}
}
\value{
data list.
}
\description{
Assign tagNum using cliqueMS's results.
}
\examples{
data <- assign_tagNum(data, plexPara = plexPara, thread = 3)
}
