% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.R
\name{.alignRt}
\alias{.alignRt}
\title{Align two chromatograms}
\usage{
.alignRt(chr1, chr2)
}
\arguments{
\item{chr1}{A XChromatogram object.}

\item{chr2}{A XChromatogram object.}
}
\value{
A list containing the aligned chr1 and chr2.
}
\description{
Align two chromatograms on retention time using approx function.
}
\examples{
chr1 <- xcms::chromPeakChromatograms(data$rawData, peaks = "CP24545")[1]
chr2 <- xcms::chromPeakChromatograms(data$rawData, peaks = "CP15906")[1]
.alignRt(chr1 = chr1, chr2 = chr2)
}
