% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tagNum.R
\name{.assign_tagNum_cliqueMS}
\alias{.assign_tagNum_cliqueMS}
\title{Assign tagNum}
\usage{
.assign_tagNum_cliqueMS(data, isoMzDiff = 0.01, thread = 1)
}
\arguments{
\item{data}{A data list.}

\item{isoMzDiff}{The mz difference between a compound and its isotope.}

\item{thread}{thread.}
}
\value{
A data list.
}
\description{
Assign tagNum using cliqueMS isotope result.
}
\examples{
data <- .assign_tagNum_cliqueMS(data, isoMzDiff = 0.01, thread = 3)
}
