% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tagNum.R
\name{.assign_tagNum_MetaboMultiplex}
\alias{.assign_tagNum_MetaboMultiplex}
\title{Assign tagNum}
\usage{
.assign_tagNum_MetaboMultiplex(
  data,
  isoMzDiff = 0.01,
  isoRtDiff = 1,
  thread = 3
)
}
\arguments{
\item{data}{A data list.}

\item{isoMzDiff}{The mz difference between a compound and its isotope.}

\item{isoRtDiff}{The rt difference between a compound and its isotope.}

\item{thread}{thread.}
}
\value{
A data list.
}
\description{
Assign tagNum using MetaboMultiplex method.
}
\examples{
data <- .assign_tagNum_MetaboMultiplex(data, thread = 3)
}
