% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.R
\name{.compare_peaks}
\alias{.compare_peaks}
\title{Compare two peaks shape}
\usage{
.compare_peaks(
  chr1,
  chr2,
  align = c("apex", "direct", "cross")[1],
  method = c("closest", "approx")[1]
)
}
\arguments{
\item{chr1}{A XChromatogram object.}

\item{chr2}{A XChromatogram object.}

\item{align}{Whether to align the vertices of two peaks.
Apex ("apex") aligns the vertices of two peaks directly.
Direct ("direct") dose not change the peak rtime dimension.
Cross uses ("cross") cross-correlation to compare the two peaks.}

\item{method}{Retention time align method.}
}
\value{
A score.
}
\description{
Compare two peaks shape using correlation.
}
\examples{
chr1 <- xcms::chromPeakChromatograms(data$rawData, peaks = "CP24545")[1]
chr2 <- xcms::chromPeakChromatograms(data$rawData, peaks = "CP15906")[1]
compare_peaks(chr1 = chr1, chr2 = chr2)
}
