% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.R
\name{.comparePeaks}
\alias{.comparePeaks}
\title{Compare two peaks' shape}
\usage{
.comparePeaks(
  chr1,
  chr2,
  method = c("pearson", "cross-correlation")[1],
  apexAlign = FALSE,
  plot = FALSE
)
}
\arguments{
\item{chr1}{A XChromatogram object.}

\item{chr2}{A XChromatogram object.}

\item{method}{Approach to calculating peak shape similarity ("pearson" and "cross-correlation").}

\item{apexAlign}{Whether to align the peaks' apex.}

\item{plot}{Whether to plot two aligned peaks.}
}
\value{
Peak shape similarity score.
}
\description{
Compare two peaks' shape using pearson or cross-correlation function.
}
\examples{
chr1 <- xcms::chromPeakChromatograms(data$rawData, peaks = "CP24545")[1]
chr2 <- xcms::chromPeakChromatograms(data$rawData, peaks = "CP15906")[1]
.comparePeaks(chr1 = chr1, chr2 = chr2, method = "cross-correlation", apexAlign = FALSE, plot = TRUE)
}
