% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportRes}
\alias{exportRes}
\title{Export Restable}
\usage{
exportRes(data, file = "resTable.xlsx")
}
\arguments{
\item{data}{data list.}

\item{file}{file path.}
}
\description{
Export Restable
}
\examples{
exportRes(data = data, file = "D:/fudan/resTable.xlsx")
}
