% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fgIdentification}
\alias{fgIdentification}
\title{Metabolite Identification}
\usage{
fgIdentification(
  data,
  cmpLibrary,
  ms2Library,
  rt_weight = 0.7,
  ms2_weight = 0.3
)
}
\arguments{
\item{data}{data list.}

\item{cmpLibrary}{cmpLibrary.}

\item{ms2Library}{ms2Library.}

\item{rt_weight}{rt_weight.}

\item{ms2_weight}{ms2_weight.}
}
\value{
A data list.
}
\description{
Metabolite identification using retention time and ms2.
}
\examples{
data <- fgIdentification(data = data, cmpLibrary = cmpLibrary, ms2Library = ms2Library)
}
