% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{getQuantRes}
\alias{getQuantRes}
\title{Get Quant Result}
\usage{
getQuantRes(data, plexPara)
}
\arguments{
\item{data}{data list.}

\item{plexPara}{plexPara.}
}
\value{
A data list.
}
\description{
Get Quant Result
}
\examples{
data <- getQuantRes(data, plexPara = plexPara)
}
