% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ms2.R
\name{getSpectra2}
\alias{getSpectra2}
\title{Assign MS2}
\usage{
getSpectra2(data, thread = 1, mzdiff = 0.001, rtdiff = 0)
}
\arguments{
\item{data}{Data list with rawData after peak annotation.}

\item{thread}{Parallel thread.}

\item{mzdiff}{c(mzmin - mzdiff, mzmax + mzdiff) constitues the precursor mz search range.
The mzdiff is determined by the number of decimal places in the spectra precursor mz recording.
Example: precursorMz = 312.83, mzdiff = 0.01.}

\item{rtdiff}{c(rtmin - rtdiff, rtmax + rtdiff) constitues the retention time search range.}
}
\value{
A data list.
}
\description{
Each peak is assigned a ms2.
}
\examples{
data <- getSpectra2(data = data)
}
