% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_library.R
\name{load_AmidoLibrary}
\alias{load_AmidoLibrary}
\title{Load AmidoLibrary}
\usage{
load_AmidoLibrary()
}
\value{
A data frame.
}
\description{
Load AmidoLibrary
}
\examples{
cmpLibrary <- load_AmidoLibrary()
}
