% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_library.R
\name{load_DEANSBANK}
\alias{load_DEANSBANK}
\title{Load DEANS-BANK}
\usage{
load_DEANSBANK(thread = 1)
}
\arguments{
\item{thread}{thread.}
}
\value{
A DataFrame object.
}
\description{
Load DEANS-BANK
}
\examples{
ms2Library <- load_DEANSBANK(thread = 3)
}
