% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_align.R
\name{peakAligning}
\alias{peakAligning}
\title{Peak Aligning}
\usage{
peakAligning(data, plexPara)
}
\arguments{
\item{data}{data list.}

\item{plexPara}{plexPara list.}
}
\value{
A data list.
}
\description{
Peak Aligning
}
\examples{
data <- peakAligning(data = data, plexPara = plexPara)
}
