% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_data.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{Peak Annotation}
\usage{
peakAnnotation(
  data,
  polarity = "positive",
  adinfo,
  ppm = 10,
  chunkSize = 1,
  thread = 1
)
}
\arguments{
\item{data}{data list with rawData after peak picking.}

\item{polarity}{positive or negative.}

\item{adinfo}{adduct information.}

\item{ppm}{Relative error in ppm to consider that two features have the mass difference of an isotope.}

\item{chunkSize}{Sample size per parallel operation.}

\item{thread}{Parallel thread.}
}
\value{
A data list.
}
\description{
Peak Annotation using cliqueMS.
}
\examples{
dataPara <- set_dataPara(data_dir = "D:/fudan/Projects/2024/MultichannelR/Progress/build_package/test_data/AP/mix1/",
res_dir = "D:/fudan/Projects/2024/MultichannelR/Progress/build_package/tmp",
sampleData = data.frame(sample_id = c("mix1_1", "mix1_2", "mix1_3"), injection_index = 1:3))
data <- load_data(dataPara = dataPara)
data <- peakPicking(data, xcmsPara = set_xcmsPara())
data(positive.adinfo, package = "cliqueMS")
positive.adinfo <- positive.adinfo[positive.adinfo$adduct \%in\%
                                     c("[M+H]+", "[M+H-H2O]+", "[M+Na]+", "[M+H-NH3]+", "[M+K]+", "[M+NH4]+"),]
data <- peakAnnotation(data, polarity = "positive", adinfo = positive.adinfo, chunkSize = 1, thread = 1)
}
