% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_group.R
\name{peakGrouping}
\alias{peakGrouping}
\title{Peak Grouping}
\usage{
peakGrouping(data, plexPara, thread = 1, extra_formula = "C14H15NO2S")
}
\arguments{
\item{data}{data list.}

\item{plexPara}{plexPara list.}

\item{thread}{thread.}

\item{extra_formula}{Additional molecular formula from derivatisation.}
}
\value{
A data list.
}
\description{
Peak grouping step.
}
\examples{
peakGrouping(data = data, plexPara = plexPara, thread = 3)
}
