% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peakGroup.R
\name{plot_peakGroup}
\alias{plot_peakGroup}
\title{Visualise a peak group}
\usage{
plot_peakGroup(
  data,
  pgid,
  type = c("chromatograms", "spectra", "cross-correlation", "pearson")[1]
)
}
\arguments{
\item{data}{data list.}

\item{pgid}{pgid.}

\item{type}{Visualisation type.}
}
\value{
No return.
}
\description{
Visualise a peak group with four types, chromatograms, spectra, cross-correlation and pearson.
}
\examples{
plot_peakGroup(data = data_AP_mix1_Amine, pgid = "PG049", type = "chromatograms")
}
