% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{set_centwavePara}
\alias{set_centwavePara}
\title{Set centwavePara}
\usage{
set_centwavePara(
  ppm = 30,
  peakwidth = c(4, 30),
  snthresh = 1,
  noise = 100,
  prefilter = c(3, 100),
  firstBaselineCheck = FALSE
)
}
\value{
A CentWaveParam object
}
\description{
See xcms::CentWaveParam function help.
}
\examples{
centwavePara <- set_centwavePara()
}
