% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{set_dataPara}
\alias{set_dataPara}
\title{Set data parameters}
\usage{
set_dataPara(data_dir, res_dir, sampleData)
}
\arguments{
\item{data_dir}{Sample directory.}

\item{res_dir}{Results directory.}

\item{sampleData}{Sample information data frame.}
}
\value{
A parameter list.
}
\description{
Set data parameters.
}
\examples{
set_dataPara(data_dir = "D:/fudan/Projects/2024/MultichannelR/Progress/build_package/test_data/AP/mix1/",
             res_dir = "D:/fudan/Projects/2024/MultichannelR/Progress/build_package/tmp",
             sampleData = data.frame(sample_id = c("mix1_1", "mix1_2", "mix1_3"), injection_index = 1:3))
}
