% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{set_xcmsPara}
\alias{set_xcmsPara}
\title{Set xcmsPara}
\usage{
set_xcmsPara(
  ppm = 30,
  peakwidth = c(4, 30),
  snthresh = 1,
  noise = 100,
  prefilter = c(3, 100),
  firstBaselineCheck = FALSE,
  mergePeak = TRUE,
  expandRt = 2,
  expandMz = 0.01,
  minProp = 0.75
)
}
\arguments{
\item{mergePeak}{Whether to merge neighbouring peaks.}
}
\value{
A xcmsPara list.
}
\description{
See xcms::CentWaveParam function help.
See xcms::MergeNeighboringPeakParam function help.
}
\examples{
xcmsPara <- set_xcmsPara()
}
