% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tagNum.R
\name{tagNumAssigning}
\alias{tagNumAssigning}
\title{Assign tagNum for peaks}
\usage{
tagNumAssigning(
  data,
  isoMzDiff = 0.01,
  isoRtDiff = 1,
  method = c("MetaboMultiplex", "cliqueMS", "cliqueMS+MetaboMultiplex",
    "MetaboMultiplex+cliqueMS")[1],
  thread = 3
)
}
\arguments{
\item{data}{A data list.}

\item{isoMzDiff}{The mz difference between a compound and its isotope.}

\item{isoRtDiff}{The rt difference between a compound and its isotope.}

\item{method}{MetaboMultiplex or cliqueMS.}

\item{thread}{thread.}
}
\value{
A data list.
}
\description{
Assign tagNum using MetaboMultiplex method or cliqueMS method.
}
\examples{
data <- tagNumAssigning(data)
}
