% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ms2.R
\name{getSpectra2}
\alias{getSpectra2}
\title{Assign MS2}
\usage{
getSpectra2(data, thread = 1, mzdiff = 0.001, rtdiff = 0)
}
\arguments{
\item{data}{data list with rawData after peak annotation.}

\item{thread}{thread.}

\item{mzdiff}{mzdiff.}

\item{rtdiff}{rtdiff.}
}
\value{
A data list.
}
\description{
Each peak is assigned a ms2.
}
\examples{
getSpectra2(data = data)
}
