% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_align.R
\name{peakAligning}
\alias{peakAligning}
\title{Peak Aligning}
\usage{
peakAligning(data, absMz = 0.02, absRt = 40)
}
\arguments{
\item{data}{data list.}

\item{absRt}{Maximum tolerated distance for retention time of peakGroups which are from same metabolite.}

\item{absMass}{Maximum tolerated distance for mass values of peakGroups which are from same metabolite.}
}
\value{
A data list.
}
\description{
Peak Aligning
}
\examples{
data <- peakAligning(data = data, plexPara = plexPara)
}
