% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_data.R
\name{peakPicking}
\alias{peakPicking}
\title{Peak picking}
\usage{
peakPicking(
  data,
  centwavePara,
  chunkSize = 3L,
  BPPARAM = BiocParallel::SnowParam(workers = 3)
)
}
\arguments{
\item{data}{data.}

\item{centwavePara}{centwavePara.}

\item{chunkSize}{chunkSize.}

\item{BPPARAM}{BPPARAM.}
}
\value{
A data list.
}
\description{
Execute the peak picking algorithm.
}
\examples{
data <- peakPicking(data, centwavePara = set_centwavePara())
}
