% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{set_plexPara}
\alias{set_plexPara}
\title{Set plexPara for program.}
\usage{
set_plexPara(
  targetGroup = c("Amine", "Phenol", "Alcohol", "Custom")[1],
  specialFrag = NULL,
  deltaMz = 2.0125,
  deltaRt = 30,
  tolMz1 = 0.01,
  tolMz2 = 0.01
)
}
\arguments{
\item{targetGroup}{Target compound group type.}

\item{specialFrag}{Special fragments from dervited metabolites.}

\item{deltaMz}{Mass difference from one peak group.}

\item{deltaRt}{Retention time difference range from one peak group.}

\item{tolMz1}{Mass difference tolerance from same fragment on ms1.}

\item{tolMz2}{Mass difference tolerance from same fragment on ms2.}
}
\value{
A parameter list.
}
\description{
Derivated metabolites produce specific secondary mass spectra which can be used to extract and identify.
}
\examples{
plexPara <- set_plexPara(targetGroup = "Amine")
}
